feat: site mirroring according to wget standards
This pull focuses on the implementation of the functionality of the `--mirror` and `--convert-links` flags.
This implementation closely follows the actual `wget` behaviour, downloading all files downloaded by the actual
program.

This is achieved by recursively traversing a websites HTML, parsing for links in tags such as `<img>`, `<a>`, etc.;
parsing the websites CSS and fetching linked resources such as background images, fonts, etc.

Not yet implemented in this release, are significantly the `-X` and `-R` options.